XP-4149 ContentTreeGrid - refactoring
Moved row formatters.
